add minimal mcp structure
