+++ libs/anaconda-assistant-mcp/LICENSE
