+++ libs/anaconda-assistant-mcp/README.md
