+++ libs/anaconda-assistant-mcp/conda.recipe/LICENSE
