+++ libs/anaconda-assistant-mcp/conda.recipe/meta.yaml
