+++ libs/anaconda-assistant-mcp/environment-dev.yml
