+++ libs/anaconda-assistant-mcp/pyproject.toml
