+++ libs/anaconda-assistant-mcp/src/anaconda_assistant_conda/cli.py
@@ -1,0 +1,23 @@
+import typer
+
+helptext = """
+The conda assistant, powered by Anaconda Assistant. \n
+See https://anaconda.github.io/assistant-sdk/ for more information.
+"""
+
+mcp_app = typer.Typer(
+    help=helptext,
+    add_help_option=True,
+    no_args_is_help=True,
+    add_completion=False,
+)
+
+
+@mcp_app.callback(invoke_without_command=True, no_args_is_help=True)
+def _() -> None:
+    pass
+
+
+@mcp_app.command(name="server-start")
+def server_start() -> None:
+    pass
